requirements.txt
fastapi[standard]
uvicorn
sqlmodel
python-dotenv
pymysql